solve: [PGM] 42576 - 완주하지못한선수 풀이
